Refactored the code
Changed the Drawable interface so that it passes
SpriteBatch and ContentMananager so that if
texture is null or the spriteBatch has changed
it can reaload itself , this may and has
bean a problem with saving and loading
a game. In the same style the draw method in player and
GameObject has been changed.
Also I changed some fields from static to normal , so we have
to create more objects, for example the explosion/